2.1-databases/work_with_database/requirements.txt
django==4.1.2
psycopg2-binary
django-autoslug~=1.9.8